+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+# v4.23
+🚀Private Release - 4.23 🚀
+MercadoPagoSDKV4 - Private Version
+- Congrats improvements
+- Handler for Biometric and ESC
+- Business Result receiptIdList and shouldShowReceipt
+
+
+# v4.22.1
+🚀Private Release - 4.22.1 🚀
+MercadoPagoSDKV4 - Private Version
+- Congrats hot fix
+
+# v4.22
+🚀Private Release - 4.22 🚀
+MercadoPagoSDKV4 - Private Version
+- Points and Discounts feature
+
 # v4.21.2
 🚀Private Release - 4.21.2 🚀
 MercadoPagoSDKV4 - Private Version